requirements.txt
djangorestframework
factory_boy
locust
pyJWT
python-decouple
pytest
pytest-benchmark
pytest-django
